requirements.txt
numpy>=1.20.0
meshio>=4.0.0
pydantic>=2.0.0
pyyaml>=6.0.0
streamlit>=1.28.0
